LayoutTests/imported/w3c/web-platform-tests/css/css-borders/border-shape/border-shape-ignore-radius-computed-expected.txt


FAIL border-radius and corner-shape are NOT reset in computed style when border-shape is specified assert_equals: expected "superellipse(0)" but got "bevel"
